feat(packages): add @jitaspace/sde-utils package with SDE download/extract/parse utilities
Extracts SDE utilities from apps/cli into a reusable package so other
consumers can download, verify, extract, and parse the EVE Online SDE
without pulling in CLI-specific dependencies.

- New package/sde-utils with: constants, downloadFile (with optional
  progress callback), sdeZipChecksum/sdeFolderChecksum/unzipSde (with
  optional progress callbacks), ensureSdePresentAndExtracted, and the
  full sdeInputFiles map + transformation helpers (loadFile, addIdToItem,
  fromArrayOfObjectsToMap, fixObjectIndices)
- apps/cli now delegates core logic to the package; cli-specific files
  become thin progress-bar wrappers, with sources/sde.ts and
  config/constants.ts reduced to re-exports
