+++ apps/cli/package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "@asteasolutions/zod-to-openapi": "^4.8.0",
+    "@jitaspace/sde-utils": "workspace:*",
     "@commander-js/extra-typings": "^14.0.0",
     "@jitaspace/eslint-config": "workspace:*",
     "@jitaspace/prettier-config": "workspace:*",
