+++ packages/sde-utils/eslint.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from "eslint/config";
+
+import { baseConfig, restrictEnvAccess } from "@jitaspace/eslint-config/base";
+
+export default defineConfig(
+  {
+    ignores: ["dist/**"],
+  },
+  baseConfig,
+  restrictEnvAccess,
+);
