+++ packages/sde-utils/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@jitaspace/sde-utils",
+  "description": "Utilities for downloading, verifying, extracting and converting the EVE Online SDE",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "license": "MIT",
+  "scripts": {
+    "clean": "rm -rf .turbo node_modules",
+    "lint": "eslint --flag unstable_native_nodejs_ts_config .",
+    "lint:fix": "pnpm lint --fix",
+    "type-check": "tsc --noEmit",
+    "prepublish": "pnpm build"
+  },
+  "dependencies": {
+    "js-yaml": "^4.1.1",
+    "node-fetch": "^2.7.0",
+    "node-stream-zip": "^1.15.0"
+  },
+  "devDependencies": {
+    "@jitaspace/eslint-config": "workspace:*",
+    "@jitaspace/prettier-config": "workspace:*",
+    "@jitaspace/tsconfig": "workspace:*",
+    "@types/js-yaml": "^4.0.9",
+    "@types/node": "^24.12.4",
+    "@types/node-fetch": "^2.6.12",
+    "eslint": "^9.39.4",
+    "prettier": "^3.8.3",
+    "typescript": "^5.9.3"
+  },
+  "prettier": "@jitaspace/prettier-config"
+}
