+++ packages/sde-utils/src/constants.ts
@@ -1,0 +1,7 @@
+export const SDE_DOWNLOAD_URL =
+  "https://developers.eveonline.com/static-data/eve-online-static-data-latest-yaml.zip";
+
+export const SDE_CHECKSUM_URL =
+  "https://eve-static-data-export.s3-eu-west-1.amazonaws.com/tranquility/checksum";
+
+export const LOCAL_SDE_FILENAME = "sde.zip";
