+++ packages/sde-utils/src/sources/sde.ts
@@ -1,0 +1,340 @@
+import fs from "fs";
+import path from "path";
+import * as YAML from "js-yaml";
+
+export type SdeSourceFile = {
+  idAttributeName: string;
+  idAttributeType: "string" | "number";
+  transformations: ((data: any, file: SdeSourceFile) => any)[];
+};
+
+export const sdeInputFiles: Record<string, SdeSourceFile> = {
+  "_sde.yaml": {
+    idAttributeName: "XXXX",
+    idAttributeType: "number",
+    transformations: [],
+  },
+  "invFlags.yaml": {
+    idAttributeName: "flagID",
+    idAttributeType: "number",
+    transformations: [fromArrayOfObjectsToMap],
+  },
+  "npcCharacters.yaml": {
+    idAttributeName: "characterID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "agentsInSpace.yaml": {
+    idAttributeName: "characterID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "agentTypes.yaml": {
+    idAttributeName: "agentTypeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "ancestries.yaml": {
+    idAttributeName: "ancestryID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "bloodlines.yaml": {
+    idAttributeName: "bloodlineID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "blueprints.yaml": {
+    idAttributeName: "blueprintTypeID",
+    idAttributeType: "number",
+    transformations: [],
+  },
+  "categories.yaml": {
+    idAttributeName: "categoryID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "certificates.yaml": {
+    idAttributeName: "certificateID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "characterAttributes.yaml": {
+    idAttributeName: "attributeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "contrabandTypes.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "controlTowerResources.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "corporationActivities.yaml": {
+    idAttributeName: "corporationActivityID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "dbuffCollections.yaml": {
+    idAttributeName: "dbuffCollectionID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "dogmaAttributeCategories.yaml": {
+    idAttributeName: "attributeCategoryID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "dogmaAttributes.yaml": {
+    idAttributeName: "attributeID",
+    idAttributeType: "number",
+    transformations: [],
+  },
+  "dogmaEffects.yaml": {
+    idAttributeName: "effectID",
+    idAttributeType: "number",
+    transformations: [],
+  },
+  "dogmaUnits.yaml": {
+    idAttributeName: "unitID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "dynamicItemAttributes.yaml": {
+    idAttributeName: "dynamicItemAttributeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "factions.yaml": {
+    idAttributeName: "factionID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "graphics.yaml": {
+    idAttributeName: "graphicID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "groups.yaml": {
+    idAttributeName: "groupID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "icons.yaml": {
+    idAttributeName: "iconID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "landmarks.yaml": {
+    idAttributeName: "landmarkID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapAsteroidBelts.yaml": {
+    idAttributeName: "asteroidBeltID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapConstellations.yaml": {
+    idAttributeName: "constellationID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapMoons.yaml": {
+    idAttributeName: "moonID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapPlanets.yaml": {
+    idAttributeName: "planetID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapRegions.yaml": {
+    idAttributeName: "regionID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapSolarSystems.yaml": {
+    idAttributeName: "solarSystemID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapStargates.yaml": {
+    idAttributeName: "stargateID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "mapStars.yaml": {
+    idAttributeName: "starID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "marketGroups.yaml": {
+    idAttributeName: "marketGroupID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "masteries.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "metaGroups.yaml": {
+    idAttributeName: "metaGroupID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "npcCorporationDivisions.yaml": {
+    idAttributeName: "npcCorporationDivisionID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "npcCorporations.yaml": {
+    idAttributeName: "corporationID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "npcStations.yaml": {
+    idAttributeName: "stationID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "planetSchematics.yaml": {
+    idAttributeName: "planetSchematicID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "races.yaml": {
+    idAttributeName: "raceID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "skinLicenses.yaml": {
+    idAttributeName: "licenseTypeID",
+    idAttributeType: "number",
+    transformations: [],
+  },
+  "skinMaterials.yaml": {
+    idAttributeName: "skinMaterialID",
+    idAttributeType: "number",
+    transformations: [],
+  },
+  "skins.yaml": {
+    idAttributeName: "skinID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "stationOperations.yaml": {
+    idAttributeName: "stationOperationID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "stationServices.yaml": {
+    idAttributeName: "stationServiceID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "translationLanguages.yaml": {
+    idAttributeName: "translationLanguageID",
+    idAttributeType: "string",
+    transformations: [],
+  },
+  "typeBonus.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "typeDogma.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "types.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "typeMaterials.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "planetResources.yaml": {
+    idAttributeName: "planetID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+  "sovereigntyUpgrades.yaml": {
+    idAttributeName: "typeID",
+    idAttributeType: "number",
+    transformations: [addIdToItem],
+  },
+};
+
+export function fromArrayOfObjectsToMap(
+  array: Record<any, any>[],
+  { idAttributeName }: SdeSourceFile,
+) {
+  const map: Record<any, any> = {};
+
+  array.forEach((item) => {
+    if (!Object.prototype.hasOwnProperty.call(item, idAttributeName)) {
+      throw new Error(`⚠️ Missing ID ${idAttributeName}`);
+    }
+    if (Object.prototype.hasOwnProperty.call(map, item[idAttributeName])) {
+      // FIXME: Downgraded to warning due to existence of ID collisions in the SDE
+      console.warn(`⚠️ Duplicate ID ${item[idAttributeName]}`);
+    }
+    map[item[idAttributeName]] = item;
+    return map[item[idAttributeName]];
+  });
+  return map;
+}
+
+export function addIdToItem(
+  obj: Record<any, any>,
+  { idAttributeName, idAttributeType }: SdeSourceFile,
+) {
+  Object.keys(obj).forEach(
+    (id) =>
+      (obj[id][idAttributeName] =
+        idAttributeType === "number" ? Number.parseInt(id, 10) : id),
+  );
+  return obj;
+}
+
+export function fixObjectIndices(
+  obj: Record<any, any>,
+  { idAttributeName }: { idAttributeName: string },
+) {
+  const result: typeof obj = {};
+  Object.values(obj).forEach(
+    (entry) => (result[entry[idAttributeName]] = entry),
+  );
+  return result;
+}
+
+export function loadFile(
+  filename: keyof typeof sdeInputFiles,
+  sdeRoot: string,
+): Record<string, any> {
+  const file = sdeInputFiles[filename];
+
+  if (!file) {
+    throw new Error(`File ${filename} not found in sdeInputFiles`);
+  }
+
+  const filePath = path.join(sdeRoot, filename);
+  let data: any = YAML.load(fs.readFileSync(filePath, "utf8"));
+
+  for (const transformation of file.transformations) {
+    data = transformation(data, file);
+  }
+
+  return data;
+}
