+++ packages/sde-utils/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "@jitaspace/tsconfig/base.json",
+  "compilerOptions": {
+    "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
+  },
+  "include": [
+    "."
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
